refactor : BigInteger를 BigDecimal로 변경

+++ src/main/java/domain/payment/CardPayment.java
@@ -1,38 +1,38 @@
 package domain.payment;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.List;
 
-import domain.menu.Category;
 import domain.OrderHistory;
+import domain.menu.Category;
 
 public class CardPayment implements PaymentStrategy{
 	@Override
-	public BigInteger pay(final List<OrderHistory> orderHistories) {
-		BigInteger paymentAmountOfChicken = calculatePaymentAmountOfChicken(orderHistories);
+	public BigDecimal pay(final List<OrderHistory> orderHistories) {
+		BigDecimal paymentAmountOfChicken = calculatePaymentAmountOfChicken(orderHistories);
 
 		return paymentAmountOfChicken.add(
 			calculatePaymentAmountOfBeverage(orderHistories)
 		);
 	}
 
-	private BigInteger calculatePaymentAmountOfChicken(final List<OrderHistory> orderHistories) {
+	private BigDecimal calculatePaymentAmountOfChicken(final List<OrderHistory> orderHistories) {
 		long chickenPaymentAmount = orderHistories.stream()
 			.filter(orderHistory -> orderHistory.isSameCategory(Category.CHICKEN))
 			.map(OrderHistory::calculatePaymentAmount)
 			.mapToLong(n -> n)
 			.sum();
 
-		return new BigInteger(String.valueOf(chickenPaymentAmount));
+		return new BigDecimal(String.valueOf(chickenPaymentAmount));
 	}
 
-	private BigInteger calculatePaymentAmountOfBeverage(final List<OrderHistory> orderHistories) {
+	private BigDecimal calculatePaymentAmountOfBeverage(final List<OrderHistory> orderHistories) {
 		long beveragePaymentAmount = orderHistories.stream()
 			.filter(orderHistory -> orderHistory.isSameCategory(Category.BEVERAGE))
 			.map(OrderHistory::calculatePaymentAmount)
 			.mapToLong(n -> n)
 			.sum();
 
-		return new BigInteger(String.valueOf(beveragePaymentAmount));
+		return new BigDecimal(String.valueOf(beveragePaymentAmount));
 	}
 }
