+++ src/main/java/domain/payment/PaymentStrategy.java
@@ -1,10 +1,11 @@
 package domain.payment;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 
 import domain.OrderHistory;
 
 public interface PaymentStrategy {
-	BigInteger pay(List<OrderHistory> orderHistories);
+	BigDecimal pay(List<OrderHistory> orderHistories);
 }
